move cluster-specific biowulf tempdir to new cluster_specific.py
