+++ workflows/rnaseq/Snakefile
@@ -8,6 +8,7 @@ import pandas as pd
 from lcdblib.snakemake import helpers, aligners
 from lcdblib.utils import utils
 from lib import common
+from lib import cluster_specific
 from lib.patterns_targets import RNASeqConfig
 
 # ----------------------------------------------------------------------------
@@ -21,7 +22,7 @@ from lib.patterns_targets import RNASeqConfig
 configfile: 'config/config.yaml'
 include: '../references/Snakefile'
 
-shell.prefix('set -euo pipefail; export TMPDIR={};'.format(common.tempdir_for_biowulf()))
+shell.prefix('set -euo pipefail; export TMPDIR={};'.format(cluster_specific.tempdir_for_biowulf()))
 shell.executable('/bin/bash')
 
 config = common.load_config(config)
